Created a simple "mock-up"/concept (with HTML, CSS and JavaScript)
This "mock-up" can serve as the base for further development.

Changes made in this commit:
- index.html
Basic markup and structure of the website using the SPA pattern. All "pages" are marked-up and organized with a <section> element for each "page", with only the "Home"-page being visible when first loading the page. The file contains a <header> for the main navigation and header-text (<h1>), a <main> element containing the previously mentioned sections, and a footer with placeholder text.
- styles.css
Basic styling of the above.
- scripts.js
Logic for the menubar on top. Clicking on menu-items will show the appropriate "page"/section and hide the rest.

Resolves #17