Implemented code which calls an API requesting random photos of foxes.
One of the main requirements in the specification for this project is to implement a button which calls a public API which returns data in the form of a JSON-object. This data is then to be presented on the page.

This commit aims to implement an experimental version of this requirement.
It contains the following implementation/changes:
- "src/index.html":
  - Added markup for displaying a random image of a fox (see notes on "src/script.js"), and a button for requesting a new image.
- "src/script.js":
  - Using AJAX, this code sends a request to this API, https://randomfox.ca/floof/, which responds with a random photo of a fox.
    - So far there is not error handling.
    - Logic for the previously mentioned button.

Resolves issue: #20