Added the page ("page-signedin.html") + logic
The content user should reach after signing in should not be loaded before they're signed in. This commit implements this page, but it is not yet used.

- "page-signedin.html":
This is the page where the data we request from the API is presentented.
  - moved the relevant markup from "./src/login.html".
- "page-signedin.js":
This is the logic which requests data from the API and updates the DOM accordingly.
  - Cleaned up the API implementation by putting all the revelant code into its own object ("getRandomFox").
  - moved the relevant logic from "./src/scripts.js".
  - The button which requests a new picture is now disabled when an image is loading, and its text is changed to "Loading..." to indicate that an image is loadng.
  -