Experiments with CSS Animations.
## "script_generate_radial-gradient_animation_keyframes":
Generates keyframes to animate "background: radial-gradient". This animates a background changing from 100% transparent to a transparent black in the form of a gradient circle with the black encroaching from the edges of the viewport and slowly filling up the viewport. See "animated-modular-popup.html" to see this in action.

## "animated-modular-popup.html":
Experimenting with css animations, e.g.
- The above-mentioned transparent black radial gradient filling up the viewport.
- Animating a modal popup window into view.

## Updated docs/resources.md